+++ aiohttp_things/__init__.py
@@ -1,4 +1,20 @@
 from aiohttp_things import views
+from aiohttp_things.views import (
+    ContextMixin,
+    ItemMixin,
+    Jinja2Mixin,
+    JSONMixin,
+    ListMixin,
+    PrimaryKeyMixin,
+)
 
 __version__ = '0.8.0'
-__all__ = ['views']
+__all__ = [
+    'views',
+    'ContextMixin',
+    'ItemMixin',
+    'Jinja2Mixin',
+    'JSONMixin',
+    'ListMixin',
+    'PrimaryKeyMixin',
+]